CA-79505: reinstate the VBD protocol negotiation
Historically, xapi used to write the protocol field into the frontend device in xenstore.  The field was based on the architecture of the VM.  This was believed to not be necessary, so was removed in the move to xenopsd.  However, some guests with buggy blkfronts, such as CentOS 5.0 and RHEL 5.0, need to be told what protocol to use.  This patch reintroduces this functionality.

Normally, we use the architecture returned by the domain builder to determine the protocol to use. In the case of an upgrade, we use the legacy domarch field in the VM object.

WARNING: This is a putative patch that has undergone no testing at all.

Signed-off-by: Jonathan Davies <jonathan.davies@citrix.com>